requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
pandas>=2.0.0
pdf2image>=1.16.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
sarvamai>=0.1.20
python-dotenv>=1.0.0
# Document preprocessing & extraction pipeline
opencv-python>=4.8.0
numpy>=1.24.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
img2table>=1.3.0
deskew>=1.5.0
paddlepaddle>=3.0.0
paddleocr>=3.0.0
